Latest Projects Section
creating a latest project section. At the moment setting up the logos and titles which will, on hover, open up a new page or fade in a new page with the projects ive been working on in that catergory! I will aim for 3 or 4 project catergories and within hopefully have 1 or 2 projects I've been working on.