project section
working on the positioning of elements within the container and also the styling. added a text section with a bit about my experience on the subject. I want to use the <br> to make a bit of a gap between the sentences i have broken. tried doing this by selecting the <br> but was unsuccessful and also tried adding a class to the breaks which was unsuccessful. I want to look into this further.