project section web development
added a video into the latest project section. At the moment it autoplays and loops. I want it to autoplay when the user opens the web development section which is my next step. I also need to do the animation so when the user hovers over the web development button it opens the corresponding section