new font & design
trying out some darker background colours and white text to break up the sections as the page was looking very 'samey'. Added a new google font to use for the titles of the project sections which looks great. Want to try and have two ections now where the web development title is over the skills section and the latets projects height spans 100% of its container div.