latest project section
created a new video for the web development section. also added a description of the work I completed.

styled the latest project section to 50vh as before it was expanding for no reason I could figure out when i added a second video to te data section. also styled some of the text sections font sizes and line heights etc.