Responsive Site
Working on the responsive css. So far I have the smaller laptop screen and tablet completed. I need to look into changing the javascript function event listeners from mouseover to click when not using a keyboard and mouse. I also want to blur out other elements when sections open on the screen.